Resolve dev merge fallout so showpullrequest CI compiles
The merge of dev into showpullrequest brought back references to types
the V1 skeleton had removed, and placed the new UC-24 tests in the wrong
package. Cleaning up:

- User.java: drop `this.inbox = new Inbox()` from the new constructor.
  Messaging is its own aggregate now (Conversation / IConversationRepository
  in messaging/), so User no longer holds an inbox field.
- CompanyManagementServiceTest: keep the real UC-24 test work from dev,
  but move it to unit/application/ where it belongs (it tests an
  application service, not a domain entity). Merged with the existing
  @Disabled placeholders for the other UCs.
- Update test fixture's Permission enum literals to match the values
  currently defined on this branch (MANAGE_INVENTORY / VIEW_SALES /
  EDIT_POLICIES); fixture data, names are not load-bearing.

mvn test: BUILD SUCCESS, 171 tests collected, 4 passing.
